Add GitHub action to refresh the database.
This commit adds a GitHub action to the database repository. When triggered, this action will check and update all of the MAP plugins in the database. This action not only allows the MAP-Client developers to streamline database maintenance, but also supports user submissions of new MAP-plugins to the database.

MAP-Client users can submit their own plugins to the database by creating an issue in the database repository. The issue should be marked with the "add-plugin" label. The issue body should contain only the GitHub repository path of the plugin to be added; this path should be in the format: {organization}/{repository}, for example: mapclient-plugins/pointsourcestep.

The action can also be triggered manually from the GitHub Actions page on the repository if we wish to update the database content without adding any new plugin repositories.

The action will also be triggered on accepting a pull request, to ensure that the content in the database remains up-to-date with the timestamp ("updated_at") on the repository.